Deploy/requirements.txt
streamlit
scikit-learn
tensorflow-cpu
pandas